add filters

+++ app/Data/Account/AccountData.php
@@ -8,7 +8,7 @@ class AccountData extends Data
 {
     public function __construct(
         public int $currency_id,
-        public int $balance,
+        public float $balance,
         public string $name,
         public string $bank,
     ) {
